Revert ffa84f2c54f03cb6851122a3015b225441532567
Thinking a bit more, this is change is not correct. The code was correct before. The BBC001 will be R1 (should be H1 but anyway). The inverted frequency scaling is already taken care of when reading the Tcal values for this BBC, since that is done by frequency. So reverting this change again.